+++ backend/src/converter/config_to_dto_conv_gen.go
@@ -47,6 +47,7 @@ func (c *ConfigToDtoConverterImpl) ConfigToSettings(source config.Config, target
 		return err
 	}
 	target.TelemetryMode = dtoTelemetryMode
+	target.LocalMaster = source.LocalMaster
 	return nil
 }
 func (c *ConfigToDtoConverterImpl) ConfigToUser(source config.Config, target *dto.User) error {