+++ CHANGELOG.md
@@ -29,8 +29,8 @@
 
 #### **🚧 Work in progress**
 
-- [ ] Manage `local master` option (?)
-- [ ] Help screen or overlay help/tour [#82](https://github.com/dianlight/srat/issues/82)
+- [x] Manage `local master` option (?)
+- [x] Help screen or overlay help/tour [#82](https://github.com/dianlight/srat/issues/82)
 - [x] Smart Control [#100](https://github.com/dianlight/srat/issues/100)
 - [x] HDD Spin down [#101](https://github.com/dianlight/srat/issues/101)